Requested edits
I removed "levels" as much as I could. Should we keep it in the title? I'm not sure what else to call it if not.

i also streamlined use of add/apply. Let me know if I missed anything.

Did a little bit of cleanup to try to simplify language.