Requested changes
I added more detail to step 4 of the bulk management section.

Since the "add roles" action isn't displaying the right theme, there isn't a way to bulk remove roles at this time. "Remove roles" used to be in the menu but it isn't anymore, so I removed that section, assuming that once the correct theme is in place that'll be resolved.